chat_api/serializers/company.py
from rest_framework import serializers

from chat_api.models import Company


class CompanySerializer(serializers.ModelSerializer):
    class Meta:
        model = Company
        fields = ["id", "name"]

