requirements.txt
django==4.1.9
djangorestframework==3.14.0
django-cors-headers==3.13.0
django-environ==0.9.0
drf-writable-nested==0.7.0

django-querycount==0.7.0
django-filter==2.4.0
django-storages==1.13.1
django-modeltranslation==0.18.7
drf-yasg==1.21.4
django-cprofile-middleware==1.0.5

psycopg[binary]==3.1.19
firebase_admin==6.0.1
requests==2.28.1

black==22.10.0
pre-commit==2.20.0

gunicorn[uvicorn]==20.1.0
uvicorn[standard]==0.20.0
python-dotenv==0.21.0

pytest==7.2.0
pytest-django==4.5.2
pytest-cov==4.0.0
model-bakery==1.9.0

sendgrid==6.9.7
google-cloud-logging==1.14.0
google-cloud-error-reporting==0.33.0
google-cloud-secret-manager==2.13.0
Pillow==9.3.0
safety==2.3.5
setuptools==67.3.1

groq==0.9.0